+++ workplace/lib/data/datasource/drive_transfer/opfs_drive_file_stager.dart
@@ -8,6 +8,7 @@ import 'package:workplace/data/datasource/drive_transfer/drive_file_stager.dart'
 import 'package:workplace/data/datasource/drive_transfer/opfs_fetch_download.dart';
 import 'package:workplace/data/datasource/drive_transfer/opfs_file_handle.dart';
 import 'package:workplace/data/datasource/drive_transfer/opfs_file_ops.dart';
+import 'package:workplace/data/datasource/drive_transfer/opfs_store.dart';
 import 'package:workplace/data/datasource/drive_transfer/staged_drive_file.dart';
 import 'package:workplace/data/model/workplace_type_defs.dart';
 import 'package:workplace/domain/entity/drive_document.dart';
